Proyecto Opcional/WindyUI/docker/StationsCronjob/app/requirements.txt
requests==2.28.2
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1